#2 몬스터가 플레이어에게 공격하는 기능 추가. 플레이어 스태미너 기능 추가.
- 플레이어는 몬스터에게 맞으면 데미지에 따라서 약경직/강경직 상태가 됨.
- 플레이어가 공격/구르기/방어 등 행동을 취하면 스태미너가 일정량 감소함. 스태미너는 초당 10씩 회복.
- 스태미너가 0이 되면 Exhausted 상태가 되며 스태미너가 80%까지 회복되기 전까지 스태미너 회복량이 감소하며 및 행동 불가 상태가 됨.

+++ TeamProject/Team Project/Assets/Scripts/Dev_이주혁님/PlayerStat.cs
@@ -44,8 +44,8 @@ public class PlayerStat
     public int curExp;    
     public int HP;
     public int curHP;
-    public int SP;
-    public int curSP;
+    public float SP;
+    public float curSP;
     public float criPro;
     public const float criMag = 1.5f;
     public int defPoint;
